+++ src/FSharp.Formatting.Literate/Document.fs
@@ -117,7 +117,7 @@ type LiterateSource =
 
 /// Representation of a literate document - the representation of Paragraphs
 /// uses an F# discriminated union type and so is best used from F#.
-type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, diagnostics) =
+type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, rootInputFolder, diagnostics) =
   // Get the content of the document as a structurally comparable tuple
   let asTuple (doc:LiterateDocument) = 
     List.ofSeq doc.DefinedLinks.Keys, List.ofSeq doc.DefinedLinks.Values,
@@ -138,6 +138,9 @@ type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, diag
   /// Location where the file was loaded from
   member _.SourceFile : string = sourceFile
 
+  /// Root for computing relative paths
+  member _.RootInputFolder : string option = rootInputFolder
+
   /// Formatted tool tips
   member _.FormattedTips : string = formattedTips
 
@@ -152,11 +155,15 @@ type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, diag
   override x.GetHashCode() = (asTuple x).GetHashCode()
 
   /// Clone the document and change some of its properties
-  member x.With(?paragraphs, ?formattedTips, ?definedLinks, ?source, ?sourceFile, ?diagnostics) =
+  member x.With(?paragraphs, ?formattedTips, ?definedLinks, ?source, ?sourceFile, ?rootInputFolder, ?diagnostics) =
     LiterateDocument
-      ( defaultArg paragraphs x.Paragraphs, defaultArg formattedTips x.FormattedTips,
-        defaultArg definedLinks x.DefinedLinks, defaultArg source x.Source, 
-        defaultArg sourceFile x.SourceFile, defaultArg diagnostics x.Diagnostics )
+      (defaultArg paragraphs x.Paragraphs,
+       defaultArg formattedTips x.FormattedTips,
+       defaultArg definedLinks x.DefinedLinks,
+       defaultArg source x.Source, 
+       defaultArg sourceFile x.SourceFile,
+       defaultArg rootInputFolder x.RootInputFolder,
+       defaultArg diagnostics x.Diagnostics )
 
 // --------------------------------------------------------------------------------------
 // Pattern matching helpers