+++ src/FSharp.Formatting.Markdown/MarkdownUtils.fs
@@ -8,10 +8,10 @@ open System.Collections.Generic
 open FSharp.Formatting.Templating
 
 module internal MarkdownUtils =
-    let isCode = (function CodeBlock(_, _, _, _, _) | InlineBlock (_, _, _) -> true | _ -> false)
+    let isCode = (function CodeBlock(_, _, _, _, _) | InlineHtmlBlock (_, _, _) -> true | _ -> false)
     let isCodeOutput = (function OutputBlock _ -> true | _ -> false)
-    let getExecutionCount = (function CodeBlock(_, executionCount, _, _, _) | InlineBlock (_, executionCount, _) -> executionCount | _ -> None)
-    let getCode = (function CodeBlock(code, _, _, _, _) -> code | InlineBlock (code, _, _) -> code | _ -> failwith "unreachable")
+    let getExecutionCount = (function CodeBlock(_, executionCount, _, _, _) | InlineHtmlBlock (_, executionCount, _) -> executionCount | _ -> None)
+    let getCode = (function CodeBlock(code, _, _, _, _) -> code | InlineHtmlBlock (code, _, _) -> code | _ -> failwith "unreachable")
     let getCodeOutput = (function OutputBlock (code, kind, _) -> code, kind | _ -> failwith "unreachable")
     let splitParagraphs paragraphs =
         let firstCode = paragraphs |> List.tryFindIndex isCode
@@ -161,7 +161,71 @@ module internal MarkdownUtils =
         let lines = code.Replace("\r\n", "\n").Split('\n') |> Array.toList
         let lines = lines |> List.filter (fun line -> line.Trim() <> sym1 && line.Trim() <> sym2 )
 
-        // Inside literate code blocks (not raw blocks) we make replacements for {{xyz}} parameters
-        let lines = lines |> List.map (SimpleTemplating.ApplySubstitutionsInText ctx.Substitutions)
         let code2 = String.concat ctx.Newline lines
         code2
+
+    let applySubstitutionsInText ctx (text: string) =
+        SimpleTemplating.ApplySubstitutionsInText ctx.Substitutions text
+
+    let rec mapSpans f (md: MarkdownSpans) =
+        md |> List.map (function
+            | Literal (text, range) -> Literal (f text, range)
+            | Strong (spans, range) -> Strong (mapSpans f spans, range)
+            | Emphasis (spans, range) -> Emphasis (mapSpans f spans, range)
+            | AnchorLink (link, range) -> AnchorLink (f link, range)
+            | DirectLink (spans, link, title, range) ->
+                DirectLink (mapSpans f spans, f link,
+                    Option.map (f) title, range)
+            | IndirectLink (spans, original, key, range) -> IndirectLink (mapSpans f spans, original, key, range)
+            | DirectImage (body, link, title, range) ->
+                DirectImage (f body, f link,
+                    Option.map (f) title, range)
+            | IndirectImage (body, original, key, range) ->
+                IndirectImage (f body, original, key, range)
+            | HardLineBreak (range) -> HardLineBreak (range) 
+
+            // NOTE: substitutions not applied to Latex math, embedded spans or inline code
+            | InlineCode (code, range) -> InlineCode (code, range)
+            | LatexInlineMath (code, range) -> LatexInlineMath (code, range) 
+            | LatexDisplayMath (code, range) -> LatexDisplayMath (code, range) 
+            | EmbedSpans (customSpans, range) -> EmbedSpans (customSpans , range)
+            )
+    let rec mapParagraphs f (md: MarkdownParagraphs) =
+        md |> List.map (function 
+        | Heading (size, body, range) ->
+           Heading (size, mapSpans f body, range)
+        | Paragraph (body, range) ->
+           Paragraph (mapSpans f body, range)
+        | CodeBlock (code, count, language, ignoredLine, range) ->
+            CodeBlock (f code, count, language, ignoredLine, range)
+        | OutputBlock (output, kind, count) ->
+            OutputBlock (output, kind, count)
+        | ListBlock (kind, items, range) ->
+            ListBlock (kind, List.map (mapParagraphs f) items, range)
+        | QuotedBlock (paragraphs, range) ->
+            QuotedBlock (mapParagraphs f paragraphs, range)
+        | Span (spans, range) ->
+            Span (mapSpans f spans, range)
+        | LatexBlock (env, body, range) ->
+            LatexBlock (env, List.map (f) body, range)
+        | HorizontalRule (character, range) ->
+            HorizontalRule (character, range)
+        | TableBlock (headers, alignments, rows, range) ->
+            TableBlock (Option.map (List.map (mapParagraphs f)) headers,
+                alignments,
+                List.map (List.map (mapParagraphs f)) rows,
+                range)
+        | OtherBlock (lines:(string * MarkdownRange) list, range) ->
+            OtherBlock (lines |> List.map (fun (line, range) -> (f line, range)), range)
+        | InlineHtmlBlock (code, count, range) -> InlineHtmlBlock (f code, count, range)
+
+        // NOTE: substitutions are not currently applied to embedded LiterateParagraph which are in any case eliminated
+        // before substitutions are applied.
+        | EmbedParagraphs (customParagraphs, range) ->
+            //let customParagraphsR = { new MarkdownEmbedParagraphs with member _.Render() = customParagraphs.Render() |> mapParagraphs f }
+            EmbedParagraphs (customParagraphs, range)
+            )
+
+    let applySubstitutionsInMarkdown ctx md =
+        mapParagraphs (applySubstitutionsInText ctx) md
+